+++ src/lib_utils/concurrency/common.rs
@@ -1,1 +1,0 @@
-pub(super) type Job = Box<dyn FnOnce() + Send + 'static>;